Rename tensor/functions/ folder to tensor/wrappers/ (Closes #55)
Folder rename. The files in this directory are AST-node wrapper classes
(`inner_product_wrapper`, `outer_product_wrapper`, `permute_indices_wrapper`,
etc.) — not free functions. The old folder name "functions" was a
misnomer that has caused confusion in code review. "wrappers" matches
the `_wrapper` suffix convention used by the class names themselves.

- `include/numsim_cas/tensor/functions/` -> `include/numsim_cas/tensor/wrappers/`
- Updated includes in `tensor_definitions.h`, `tensor_std.h`,
  `skew_classification.h`, `projector_algebra.h`
- Updated doc reference in `docs/simplifier-coverage.md`

Done with `git mv` so file history is preserved. No behavior change.
All 1497 tests pass.

Stacks on #52 (class rename `basis_change_imp` -> `permute_indices_wrapper`).
