+++ include/numsim_cas/tensor/wrappers/inner_product_wrapper.h
