+++ include/numsim_cas/tensor/wrappers/outer_product_wrapper.h
