+++ include/numsim_cas/tensor/wrappers/simple_outer_product.h
