+++ include/numsim_cas/tensor/wrappers/tensor_pow.h
